+++ playwright-bdd/package.json
@@ -3,15 +3,17 @@
     "version": "1.0.0",
     "main": "index.js",
     "scripts": {
-        "bdd:login": "bddgen && playwright test tests/bdd/features/innlogging-alle-flater.feature --project=bdd",
-        "bdd:refresh": "bddgen && playwright test tests/bdd/features/innlogging-refresh-alle-flater.feature --project=bdd",
-        "bdd:logout": "bddgen && playwright test tests/bdd/features/utlogging-alle-flater.feature --project=bdd",
-        "bdd:infoportal": "bddgen && playwright test tests/bdd/features/infoportal-innlogging.feature --project=bdd",
-        "bdd": "bddgen && playwright test tests/bdd/features/*.feature --project=bdd",
-        "native:login": "playwright test tests/native/test-cases/innlogging-alle-flater.spec.ts --project=native",
-        "native:refresh": "playwright test tests/native/test-cases/innlogging-refresh-alle-flater.spec.ts --project=native",
-        "native:logout": "playwright test tests/native/test-cases/utlogging-alle-flater.spec.ts --project=native",
-        "native:infoportal": "playwright test tests/native/test-cases/infoportal-innlogging.spec.ts --project=native",
+        "bdd:sso:login": "bddgen && playwright test tests/bdd/features/sso/innlogging-alle-flater.feature --project=bdd",
+        "bdd:sso:refresh": "bddgen && playwright test tests/bdd/features/sso/innlogging-refresh-alle-flater.feature --project=bdd",
+        "bdd:sso:logout": "bddgen && playwright test tests/bdd/features/sso/utlogging-alle-flater.feature --project=bdd",
+        "bdd:sso:infoportal": "bddgen && playwright test tests/bdd/features/sso/infoportal-innlogging.feature --project=bdd",
+        "bdd:sso": "bddgen && playwright test tests/bdd/features/sso/*.feature --project=bdd",
+        "bdd": "bddgen && playwright test tests/bdd/features/**/*.feature --project=bdd",
+        "native:sso:login": "playwright test tests/native/test-cases/sso/innlogging-alle-flater.spec.ts --project=native",
+        "native:sso:refresh": "playwright test tests/native/test-cases/sso/innlogging-refresh-alle-flater.spec.ts --project=native",
+        "native:sso:logout": "playwright test tests/native/test-cases/sso/utlogging-alle-flater.spec.ts --project=native",
+        "native:sso:infoportal": "playwright test tests/native/test-cases/sso/infoportal-innlogging.spec.ts --project=native",
+        "native:sso": "playwright test tests/native/test-cases/sso/*.spec.ts --project=native",
         "native": "playwright test --project=native"
     },
     "keywords": [],
